CASMPET-4185: Add checks for K8s etcd not configured correctly
We had a deploy failure on drax that took a long time to figure
out that the problem was that the etcd instance on a master wasn't
healthy.

The problem seen was that charts failed to install and this led to
seeing that the API server was unhealthy and this finally led to
looking at the etcd instances on the different masters and that
the cloud-init wound up messing up the etcd config after it was
already running. The indications that there was a problem weren't
that the etcd service was running (it was), but that there were
missing files and also that an attempt to run the
`etcdctl endpoint health` command failed. This is described in
CASMINST-2265.

This change adds a goss test that would catch this condition by
checking for the files that should be there and verifying that
`etcdctl endpoint health` works.