CASMPET-4187: Run etcd-service test during upgrade
While there's already a test during upgrade to check that the
etcd cluster is seeing 3 members, it might be useful to run the
etcd-service test which tests things related to the local instance:
it's running and, given [0] it checks if the cert files are in
place and if the endpoint health command works. This should help
to point to which node has the problem if there is one.

[0] https://stash.us.cray.com/projects/CSM/repos/csm-testing/pull-requests/169/diff