CASMINST-3417: remove bogus curl call
The original author of this line of code can't recall its purpose,
and it has no obvious discernible purpose on the face of it. Further,
it fails when invoked because this reaches out to nexus, not arti,
and nexus has no idea what to do with that attempted endpoint/api
reference, generating obnoxious error/warning spew. This commit
removes the mystery curl.