CASMTRIAGE-2343 fix bond link test
this test runs in two contexts.  on the livecd, the bond config files have the
bond members in single quoutes, but on the ncns, they are just text.  this
commit just adds an additional piped command that strips the single quotes from
the output, resulting in just the bond member names.  this allows the while loop
to function properly, thus the test.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>