CASMINST-5088: Scripts to validate kyverno configuration. (#376)
* Script to check number of kyverno running pods

This script checks for Kyverno pods in running state. If any issue is encountered, we can run this script manually to check appropriate error message which helps in addressing the issue. This script will be part of goss testing framework in next release.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Script to check kyverno policy report

This script checks for Kyverno policy report details. If any failures/warnings/erros/skipped count in the report, the test fails. We can run this script manually to check appropriate error message which helps in addressing the issue. This script will be part of goss testing framework in next release.

Signed-off-by: pradeepkumargl <80017325+pradeepkumargl@users.noreply.github.com>

* Linting

Signed-off-by: Mitch Harding <mitchell.harding@hpe.com>
Co-authored-by: Mitch Harding <mitchell.harding@hpe.com>