MTL-2088 Fix build race condition (#439)
Variables need a `def` in order to be unique/re-defined
within each `matrix` build, otherwise there is a risk of
a race condition happening. The race condition occurs
when one `matrix` build redefines (reuses) the variables
while the variables are being read by another `matrix` build.

In MTL-2088, an SP4 build was publishing RPMs when the SP3
`matrix` build updated these variables, leading to some of
the SP4 RPMs to end up in the SP3 drop-off.