goss-servers is missing goss requirement
The goss-servers package needs goss to function, but this requirement is not specified in the spec.

This has caused a build failure where we removed the goss package, and we (I) expected goss-servers to find either hpe-csm-goss-package or goss depending on which one was available for the respective architecture (aarch64 or x86_64).