CASMTRIAGE-7627 ignore non-csm failed backups
`velero_backups_check.sh` causes a test failure intermittently on
in-house sytems, often from other products such as PBS or Slingshot.
While they may be valid failures, the CSM product is not at fault, so
those should not trigger a test failure.

for this commit, I modified it to check only for failed vault backups
since that is more pertinent to CSM itself.

I moved the code that was previously run with a pipe and checking the
exit code to instead make a temp file, use a single jq command to query
it, and then fail if there are any failed vault backups.

if we determine other backups should be checked, we can add those to
    that jq command.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>
(cherry picked from commit 164e86dcdc31020efbaeb380486d6fa8a8d58c18)
Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>