CASMTRIAGE-7657 adjust k8s_verify_cluster_2 test to avoid some false positives by checking if a newer pod was successful.
the test currently fails when any output is detected.  it does so when
any error pod is detected.  however, there are cases when one pod fails,
but another is started and is then successful.

this commit adds extra logic to check if a newer pod exists and is a
non-error state, which is then presumed to be acceptable.  no output is
printed in this case.

otherwise, the podname is printed triggering the test to fail.

it was necessary to move the code to an external script because the
logic and special characters were proving too cumbersome to handle in
the limited `exec:` block

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>
(cherry picked from commit c1c0bbbdf4e3f915328862a10bdfada5de5efe61)