CASMINST-7168 adjust `velero_backups_check.sh` to ignore failed (#677)
backups if a newer successful one exists

during nightly CT runs, velero backups for vault have failed, but
newer successful backups existed.  the original issue was transient
and due to a timeout error, per CASMTRIAGE-7762, but the ct tests
continued to generate auto-triage tickets since a failure was detected.

by extending this script to check for the existence of a newer
successful backup and not failing in that case, we will prevent
additional, irrelevant auto-triage tickets from being generated.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>