CASMINST-7340 temporarily exclude slurmdbd-backup pod from goss test until USS-3925 is resolved
USS-3925 is a known-issue where the slurmdbd-backup pod has 2/3 running
containers, which can cause this test to fail.  until that is resolved,
the existing test will always report failures.

the ansible play checks if the deployment is ready before proceeding, so
there is a level of assurance the cluster is deployed.  but to prevent
false tickets when this is in the pipeline, this test temporarily skips
that pod.

this also fixes a missing `{` in a test title

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>