DST-14865 move csm-testing-internal to internal repo
due to some limitations in the pipeline, DST cannot pull from algol60
so I am moving all of this to an internal repo.  this restores the repo
to the state it was prior to the commits I added for this and it will
now only make the goss-servers and csm-testing rpms as it did not too
long ago.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>