CRAYSAT-2005: Improve `sat bootprep` skip/overwrite tests (#721)
Improve testing for the skip and overwrite functionality of sat bootprep.

Add tests for the if_exists property with three new bootprep input
files, one for CFS configs, one for IMS images, and one for session
templates. Each one defines two items, one that specifies skip and one
that specifies overwrite for the value of if_exists.

Modify the test case that creates a CFS configuration, IMS image, and
BOS session template to exercise the --skip-* and --overwrite-*
options instead of specifying if_exists in the input file.

Add some functionality to create empty IMS images to speed up the
testing of image overwrites. This will be further improved in
CRAYSAT-2008.

Remove the if_exists property from most of the other bootprep input
files, so that we test when that property is not specified, which is
probably the most likely usage of the command today. This was originally
added to ensure that the tests would work without pausing for
interactive prompts, but as long as our cleanup logic is solid, this
should not be necessary. Cleanup will be further improved in
CRAYSAT-2012.

Increase timeout of `sat bootprep` tests in goss file from 20 minutes to
45 minutes. On rocket, tests took about 35 minutes.

Minor other cleanup inluding:
- Move `create_vcs_repository` call out of `get_csm_vars` and into
  `setUpClass`
- Use `self.assertNotIn` instead of `self.assertFalse` in a few spots
- Split lines greater than 120 chars
- Properly align line continuations
- Fix a docstring to follow google style

Test Description:

Executed tests with the sat_functional binary on rocket. All tests
passed.