CRAYSAT-2018: Fix duplicate Goss results for SAT tests (#731)
The usage of both `stderr` and `exit-status` checks in the Goss files
for the SAT functional tests resulted in duplicate results for each
test. Remove the `stderr` check as `exit-status` is sufficient to check
for success or failure of the test.

Test Description:
Made this change to a single Goss test file on noname and executed the
goss test with `goss -g tests/goss-sat-version.yaml validate --format
json` and confirmed that only a single test result was returned.