CRAYSAT-2017: Skip tests when required CSM data is missing (#733)
Add decorators to each test method that allows the test to be skipped if
required data from the product catalog about the installed CSM versions
is missing.

For example, tests that require the latest CSM version to have a
barebones recipe check whether such a recipe is found in the csm_vars
set up by `get_csm_vars`.

Sometimes the tests use bootprep files that directly reference these
variables, and other times the tests are checking the presence of the
variable as a proxy to check for the presence of the required data in
the product catalog. For example, a test may check for `csm.recipe_id`,
and it's bootprep file may not reference that variable, but it checks
this variable to ensure a barebones recipe should be found by bootprep.

Improve the skip message when there are missing variables by only
printing the list of csm vars. The `skipTest` method does not nicely
print multi-line strings and `set_up_errors` are already logged.

Test Description:
Executed on wasp where the failure was originally observed.

Modified the `test_bootprep` module to look for an additional key in the
CSM versions on the system to simulate what it should do when variables
are not defined. Confirmed tests were skipped with a reasonable skip
message about the missing variables.