CRAYSAT-2013: Improve `sat hwinv` summary tests (#740)
Improve the tests of the `sat hwinv --summarize*` options by defining a
new method `assert_summary_sections` to reduce the amount of duplicated
code across the tests. All the tests were checking for the same things,
with variability in the component type name, expected summary keys, and
whether a listing was expected for the components. These variables are
captured as arguments in the new `assert_summary_sections` method.

This method is used across all the other `test_` methods with the
appropriate arguments to make the same assertions as before.

The default list of summary fields for each component type is also saved
in teh `setUp` method to remove the need to duplicate that across
multiple test methods.

Rename the existing `TestHwinv` class to `TestHwinvList`, and move
summarize tests to `TestHwinvSummarize`. These tests can be run in
parallel.

Test Description:
Not tested yet.