CRAYSAT-2035: Clean up deleted images after `sat bootprep` tests
The `sat bootprep` functional tests perform tests that check that `sat
bootprep` will overwrite existing images. When `sat bootprep` does that,
it deletes the old image, but does not permanently delete it, so it
remains in IMS `/deleted/images`. Clean this up in the `tearDown` of the
`BootprepTestCase` class so that we don't leave that cruft on the
system, which will accumulate over time as these tests execute.

Test Description:
Extracted testing RPM and ran with Goss on vinland. Verified that no
artifacts remained on the system that started with the "sat-test" prefix
after tests had finished executing.