CRAYSAT-2006: Add 'sat bootprep' test for image_ref
Modify the `configs-images-and-session-templates.yaml` bootprep input
file used by `test_configs_images_and_session_templates` to include
multiple images where one image depends on another image and refers to
it by its `ref_name` with the `base.image_ref` property. The first image
is built from a recipe, and the second is configured from that resulting
image.  In addition, update the session template to refer to the image
by its `ref_name` with `image.image_ref`. This is important to test
because this is how our default bootprep input files that get delivered
with the HPC CSM software recipe are set up.

Modify `test_dry_run_and_save` to check for the right log messages being
logged now that there are two images in the input file.

Modify `test_limit_option` to use a separate input file that uses a
simpler input file that only defines one config, one image, and one
session template, none of which use `image_ref` and `ref_name`. It does
not work to use `image_ref` to refer to an image in the input file if
you use `--limit session_templates` because bootprep does not know the
image ID for the referenced image if it was not created by `bootprep`.
This is arguably a bug in `bootprep`, but it is not new, and it is not a
critical use case, fortunately. So just fix the test to not try to do
this. Additionally use `--dry-run` on the image use case to reduce the
amount of time spent building images in the tests.

Test Description:
Copied test RPM to vinland, extracted, and ran Goss tests.