CRAYSAT-2034: Skip bootprep image build tests if repos unavailable (#749)
If the SLE products are not installed on a system, the repositories used
by the CSM barebones image are not available on the system. Add a check
for whether those products are installed on the system, and skip the
`sat bootprep` tests that build images from the barebones recipe if the
SLE products are not installed.

This code was largely generated by GitHub Copilot.

Test Description:
Tested on wasp, which only has CSM 1.7.0 installed and on the vshasta
system beau, which also only has CSM 1.7.0 installed.